Room Database and dynamic links integration

+++ app/src/main/java/com/easy/meet/MainActivity.kt
@@ -47,7 +47,7 @@ class MainActivity : ComponentActivity() {
 
                     when {
                         uri.toString().contains("event") -> {
-                            val eventID = path.toInt()
+                            val eventID = path.toLong()
                             Log.e("Aditi===>", "event ID from dynamic link::$eventID")
 
                         }
